COreções de repositórios para não usarem native queries (Sonar)

+++ src/main/java/br/com/basis/abaco/repository/UserRepository.java
@@ -10,6 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigInteger;
 import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -52,7 +53,7 @@ public interface UserRepository extends JpaRepository<User, Long> {
     Page<User> findAllByLoginNot(Pageable pageable, String login);
 
     @Query(value = "select tipo_equipe_id from user_tipo_equipe where user_id = :idUser", nativeQuery = true)
-    List<Long> findUserEquipes(@Param("idUser") Long idUser);
+    List<BigInteger> findUserEquipes(@Param("idUser") Long idUser);
 
     @EntityGraph(attributePaths = {"authorities","tipoEquipes","organizacoes"})
     @Query(value = "select u from User u join fetch u.organizacoes o where u.login=?1 and u.activated=true and o.ativo=true")
